Added Usage to README. Switching to unicode strings. Other minor changes.
'encoding' flag added to Core's __init__ to change the unicode encoding
(default utf-8)

Sockets are now (re)created on connect, not on object creation. This
means that it is no longer necessary to instantiate a new IRC object
when trying to reconnect. Added _createSocket() for this.

Some minor cleanups in code and documentation.
blackbox should be forward compatible to py3 now.